+++ src/views/form.blade.php
@@ -1,9 +1,10 @@
 <form {{ $formAttributes }} novalidate>
 @foreach($fields as $field)
-      {{ $field->generate() }}
+    {{ $field->generate() }}
 @endforeach
     <div class="form-actions">
-        <button type="submit" class="btn btn-primary">Save changes</button>
-        <button type="reset" class="btn">Cancel</button>
+@foreach($actions as $action)
+    {{ $action->generate() }}
+@endforeach
     </div>
 </form>