+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change log
 
+## 25th April 2013
+
+* Added the $form property. This defines the form element tag
+* Added actions. These are fields that will appear in the form footer. Like submit and cancel buttons
+
 ## 24th April 2013
 
 * Password fields will now never include the value attribute in forms for secrity reasons.